# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Система бронирования номеров",
        "contact": {}
    },
    "paths": {
...
        "/bookings/{id}/pay": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает платеж через YooKassa для указанного бронирования и возвращает ссылку для оплаты.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Создание платежа для бронирования",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ссылка для оплаты успешно создана",
                        "schema": {
                            "$ref": "#/definitions/response.CreatePaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос или бронирование уже оплачено",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Бронирование не найдено",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера или ошибка платежной системы",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bookings/{id}/refund": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обрабатывает запрос на возврат платежа за бронирование. Проверяет права доступа пользователя, статус оплаты и выполняет возврат через API YooKassa.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Обработка возврата платежа",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Оплата отменена",
                        "schema": {
                            "$ref": "#/definitions/response.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Бронирование не оплачено или ID платежа отсутствует",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "У вас нет прав на отмену этого бронирования",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Бронирование не найдено",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/payments/callback": {
            "post": {
                "description": "Обрабатывает уведомления от платежной системы и обновляет статус оплаты для указанного бронирования.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Webhook для обработки статуса оплаты",
                "parameters": [
                    {
                        "description": "Данные вебхука от платежной системы",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payments.PaymentCallbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус оплаты обновлен",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные запроса",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Бронирование не найдено",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "payments.PaymentCallbackRequest": {
            "type": "object",
            "properties": {
                "object": {
                    "description": "Основной объект данных",
                    "allOf": [
                        {
                            "$ref": "#/definitions/payments.PaymentObject"
                        }
                    ]
                }
            }
        },
        "payments.PaymentMetadata": {
            "type": "object",
            "properties": {
                "booking_id": {
                    "description": "Уникальный идентификатор бронирования",
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "payments.PaymentObject": {
            "type": "object",
            "properties": {
                "metadata": {
                    "description": "Метаданные оплаты",
                    "allOf": [
                        {
                            "$ref": "#/definitions/payments.PaymentMetadata"
                        }
                    ]
                },
                "status": {
                    "description": "Статус оплаты",
                    "type": "string",
                    "example": "succeeded"
                }
            }
        },
...
        "response.CreatePaymentResponse": {
            "type": "object",
            "properties": {
                "payment_url": {
                    "description": "Ссылка для оплаты",
                    "type": "string",
                    "example": "ссылка на оплату"
                }
            }
        },
        "response.ErrorResponse": {
            "description": "Стандартный ответ при ошибке",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "response.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "response.SuccessResponse": {
            "description": "Стандартный ответ при успешном выполнении",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}